Fixes issue 7307 Validate.database should be able to run for a single module and only check module-specific db artifacts Validate module is now run as part of export.database Fixes issue 6518 Export Client does not work
--HG--
extra : convert_revision : svn%3A9509db7c-3d11-0410-ba00-cf6d01345df7/trunk%4012700

+++ src-test/org/openbravo/test/dal/DalPerformanceInventoryLineTest.java
@@ -65,7 +65,7 @@ public class DalPerformanceInventoryLineTest extends BaseTest {
       for (int j = 0; j < NO_LINE; j++) {
         final InventoryCountLine icl = (InventoryCountLine) DalUtil.copy(baseLine, false);
         icl.setInventoryCount(ic);
-        icl.setLine(j);
+        icl.setLine((long) j);
         ic.getMaterialMgmtInventoryCountLineList().add(icl);
       }
       OBDal.getInstance().save(ic);
@@ -98,7 +98,7 @@ public class DalPerformanceInventoryLineTest extends BaseTest {
       cnt++;
       final InventoryCountLine icl = (InventoryCountLine) DalUtil.copy(baseLine, false);
       icl.setInventoryCount(ic);
-      icl.setLine(ic.getMaterialMgmtInventoryCountLineList().size() + 1);
+      icl.setLine((long) (ic.getMaterialMgmtInventoryCountLineList().size() + 1));
       ic.getMaterialMgmtInventoryCountLineList().add(icl);
 
       cntLine = ic.getMaterialMgmtInventoryCountLineList().size();
